Docs: 프로필 카드 UI 변경
