+++ src/components/mypage/reviewable-card/index.tsx
